+++ tests/src/Command/CoverageCheckCommandTest.php
@@ -21,7 +21,6 @@ use Esi\CoverageCheck\CoverageCheck;
 use Esi\CoverageCheck\Style\CoverageCheckStyle;
 use InvalidArgumentException;
 use PHPUnit\Framework\Attributes\CoversClass;
-use PHPUnit\Framework\Attributes\UsesClass;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Output\Output;
@@ -29,6 +28,8 @@ use Symfony\Component\Console\Tester\ApplicationTester;
 use Symfony\Component\Console\Tester\CommandTester;
 
 use function dirname;
+use function preg_replace;
+use function str_replace;
 use function trim;
 
 /**
@@ -37,12 +38,14 @@ use function trim;
 #[CoversClass(CoverageCheckCommand::class)]
 #[CoversClass(Application::class)]
 #[CoversClass(CoverageCheckStyle::class)]
-#[UsesClass(CoverageCheck::class)]
+#[CoversClass(CoverageCheck::class)]
 class CoverageCheckCommandTest extends TestCase
 {
     protected Application $application;
+
     protected ApplicationTester $tester;
-    private static $fixtures;
+
+    private static array $fixtures;
 
     protected function setUp(): void
     {
@@ -53,15 +56,17 @@ class CoverageCheckCommandTest extends TestCase
             'invalid_root' => dirname(__FILE__, 3) . '/fixtures/invalid_root_element.xml',
             'no_children'  => dirname(__FILE__, 3) . '/fixtures/no_children.xml',
             'no_metrics'   => dirname(__FILE__, 3) . '/fixtures/no_project_metrics.xml',
+            'thisLibrary'  => dirname(__FILE__, 3) . '/fixtures/self_clover.xml',
         ];
 
-        $command     = new CoverageCheckCommand(new CoverageCheck());
-        $commandName = $command->getName();
+        $coverageCheckCommand = new CoverageCheckCommand(new CoverageCheck());
+        $commandName          = $coverageCheckCommand->getName();
 
         $this->application = new Application('PHPUnit Coverage Check', CoverageCheck::VERSION);
         $this->application->setAutoExit(false);
-        $this->application->add($command);
+        $this->application->add($coverageCheckCommand);
         $this->application->setDefaultCommand($commandName, true);
+
         $this->tester = new ApplicationTester($this->application);
     }
 
@@ -73,13 +78,13 @@ class CoverageCheckCommandTest extends TestCase
         ]);
 
         self::assertEquals(self::$fixtures['invalid_root'], $this->tester->getInput()->getArgument('cloverfile'));
-        self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
+        self::assertSame(90, $this->tester->getInput()->getArgument('threshold'));
 
-        self::assertEquals(
+        self::assertSame(
             '[ERROR] Clover file appears to be invalid. Are you sure this is a PHPUnit generated clover report?',
             self::stripWhitespace($this->tester->getDisplay())
         );
-        self::assertEquals(Command::INVALID, $this->tester->getStatusCode());
+        self::assertSame(Command::INVALID, $this->tester->getStatusCode());
     }
 
     public function testCloverFileNoChildren(): void
@@ -90,13 +95,13 @@ class CoverageCheckCommandTest extends TestCase
         ]);
 
         self::assertEquals(self::$fixtures['no_children'], $this->tester->getInput()->getArgument('cloverfile'));
-        self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
+        self::assertSame(90, $this->tester->getInput()->getArgument('threshold'));
 
-        self::assertEquals(
+        self::assertSame(
             '[ERROR] Clover file appears to be invalid. Are you sure this is a PHPUnit generated clover report?',
             self::stripWhitespace($this->tester->getDisplay())
         );
-        self::assertEquals(Command::INVALID, $this->tester->getStatusCode());
+        self::assertSame(Command::INVALID, $this->tester->getStatusCode());
     }
 
     public function testCloverFileNoProjectMetrics(): void
@@ -107,21 +112,21 @@ class CoverageCheckCommandTest extends TestCase
         ]);
 
         self::assertEquals(self::$fixtures['no_metrics'], $this->tester->getInput()->getArgument('cloverfile'));
-        self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
+        self::assertSame(90, $this->tester->getInput()->getArgument('threshold'));
 
-        self::assertEquals(
+        self::assertSame(
             '[ERROR] Clover file appears to be invalid. Are you sure this is a PHPUnit generated clover report?',
             self::stripWhitespace($this->tester->getDisplay())
         );
-        self::assertEquals(Command::INVALID, $this->tester->getStatusCode());
+        self::assertSame(Command::INVALID, $this->tester->getStatusCode());
     }
 
     public function testRunInvalidCloverFile(): void
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessageMatches('/Invalid input file provided. Was given: (.*?)clovr.xml/');
-        $tester = new CommandTester($this->application->find('coverage:check'));
-        $tester->execute([
+        $commandTester = new CommandTester($this->application->find('coverage:check'));
+        $commandTester->execute([
             'cloverfile' => self::$fixtures['notexist'],
             'threshold'  => 90,
         ]);
@@ -131,8 +136,8 @@ class CoverageCheckCommandTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('The threshold must be a minimum of 1 and a maximum of 100, 101 given');
-        $tester = new CommandTester($this->application->find('coverage:check'));
-        $tester->execute([
+        $commandTester = new CommandTester($this->application->find('coverage:check'));
+        $commandTester->execute([
             'cloverfile' => self::$fixtures['valid'],
             'threshold'  => 101,
         ]);
@@ -142,8 +147,8 @@ class CoverageCheckCommandTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('The threshold must be a minimum of 1 and a maximum of 100, 0 given');
-        $tester = new CommandTester($this->application->find('coverage:check'));
-        $tester->execute([
+        $commandTester = new CommandTester($this->application->find('coverage:check'));
+        $commandTester->execute([
             'cloverfile' => self::$fixtures['valid'],
             'threshold'  => 0,
         ]);
@@ -157,13 +162,13 @@ class CoverageCheckCommandTest extends TestCase
         ]);
 
         self::assertEquals(self::$fixtures['empty'], $this->tester->getInput()->getArgument('cloverfile'));
-        self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
+        self::assertSame(90, $this->tester->getInput()->getArgument('threshold'));
 
-        self::assertEquals(
+        self::assertSame(
             '[ERROR] Insufficient data for calculation. Please add more code.',
             trim($this->tester->getDisplay())
         );
-        self::assertEquals(Command::FAILURE, $this->tester->getStatusCode());
+        self::assertSame(Command::FAILURE, $this->tester->getStatusCode());
     }
 
     public function testRunNotEnoughCodePercentageOnly(): void
@@ -175,14 +180,14 @@ class CoverageCheckCommandTest extends TestCase
         ], ['verbosity' => Output::VERBOSITY_NORMAL]);
 
         self::assertEquals(self::$fixtures['empty'], $this->tester->getInput()->getArgument('cloverfile'));
-        self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
+        self::assertSame(90, $this->tester->getInput()->getArgument('threshold'));
         self::assertTrue($this->tester->getInput()->getOption('only-percentage'));
 
-        self::assertEquals(
+        self::assertSame(
             '[ERROR] Insufficient data for calculation. Please add more code.',
             trim($this->tester->getDisplay())
         );
-        self::assertEquals(Command::FAILURE, $this->tester->getStatusCode());
+        self::assertSame(Command::FAILURE, $this->tester->getStatusCode());
     }
 
     public function testRunValidNonPassingOptions(): void
@@ -193,13 +198,13 @@ class CoverageCheckCommandTest extends TestCase
         ]);
 
         self::assertEquals(self::$fixtures['valid'], $this->tester->getInput()->getArgument('cloverfile'));
-        self::assertEquals(100, $this->tester->getInput()->getArgument('threshold'));
+        self::assertSame(100, $this->tester->getInput()->getArgument('threshold'));
 
-        self::assertEquals(
+        self::assertSame(
             '[ERROR] Total code coverage is 90.32% which is below the accepted 100%',
             trim($this->tester->getDisplay())
         );
-        self::assertEquals(Command::FAILURE, $this->tester->getStatusCode());
+        self::assertSame(Command::FAILURE, $this->tester->getStatusCode());
     }
 
     public function testRunValidOptionsNonPassingPercentageOnly(): void
@@ -211,14 +216,14 @@ class CoverageCheckCommandTest extends TestCase
         ]);
 
         self::assertEquals(self::$fixtures['valid'], $this->tester->getInput()->getArgument('cloverfile'));
-        self::assertEquals(100, $this->tester->getInput()->getArgument('threshold'));
+        self::assertSame(100, $this->tester->getInput()->getArgument('threshold'));
         self::assertTrue($this->tester->getInput()->getOption('only-percentage'));
 
-        self::assertEquals(
+        self::assertSame(
             '90.32%',
             trim($this->tester->getDisplay())
         );
-        self::assertEquals(Command::FAILURE, $this->tester->getStatusCode());
+        self::assertSame(Command::FAILURE, $this->tester->getStatusCode());
     }
 
     public function testRunValidOptionsPassing(): void
@@ -229,13 +234,13 @@ class CoverageCheckCommandTest extends TestCase
         ]);
 
         self::assertEquals(self::$fixtures['valid'], $this->tester->getInput()->getArgument('cloverfile'));
-        self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
+        self::assertSame(90, $this->tester->getInput()->getArgument('threshold'));
 
-        self::assertEquals(
+        self::assertSame(
             '[OK] Total code coverage is 90.32%',
             trim($this->tester->getDisplay())
         );
-        self::assertEquals(Command::SUCCESS, $this->tester->getStatusCode());
+        self::assertSame(Command::SUCCESS, $this->tester->getStatusCode());
     }
 
     public function testRunValidOptionsPassingPercentageOnly(): void
@@ -247,14 +252,84 @@ class CoverageCheckCommandTest extends TestCase
         ]);
 
         self::assertEquals(self::$fixtures['valid'], $this->tester->getInput()->getArgument('cloverfile'));
-        self::assertEquals(90, $this->tester->getInput()->getArgument('threshold'));
+        self::assertSame(90, $this->tester->getInput()->getArgument('threshold'));
         self::assertTrue($this->tester->getInput()->getOption('only-percentage'));
 
-        self::assertEquals(
+        self::assertSame(
             '90.32%',
             trim($this->tester->getDisplay())
         );
-        self::assertEquals(Command::SUCCESS, $this->tester->getStatusCode());
+        self::assertSame(Command::SUCCESS, $this->tester->getStatusCode());
+    }
+
+    public function testShowFilesTableOutputAboveThreshold(): void
+    {
+        $this->tester->run([
+            'cloverfile'   => self::$fixtures['thisLibrary'],
+            'threshold'    => 90,
+            '--show-files' => true,
+        ]);
+
+        self::assertEquals(self::$fixtures['thisLibrary'], $this->tester->getInput()->getArgument('cloverfile'));
+        self::assertSame(90, $this->tester->getInput()->getArgument('threshold'));
+
+        $eol = \PHP_EOL;
+
+        $expected = '------------------------------------------------------------------- --------------------------------- ---------- ' . $eol;
+        $expected .= '  File                                                                Covered Elements/Total Elements   Coverage  ' . $eol;
+        $expected .= ' ------------------------------------------------------------------- --------------------------------- ---------- ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Application.php                    10/10                             100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Command\CoverageCheckCommand.php   77/77                             100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\CoverageCheck.php                  63/63                             100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Style\CoverageCheckStyle.php       4/4                               100.00%   ' . $eol;
+        $expected .= '  [...]\phpunit-coverage-check\src\Utils.php                          16/16                             100.00%   ' . $eol;
+        $expected .= ' ------------------------------------------------------------------- --------------------------------- ---------- ' . $eol;
+        $expected .= '  Overall Totals                                                      170/170                           100.00%   ' . $eol;
+        $expected .= ' ------------------------------------------------------------------- --------------------------------- ----------';
+
+        self::assertEquals($expected, trim($this->tester->getDisplay()));
+        self::assertSame(Command::SUCCESS, $this->tester->getStatusCode());
+    }
+
+    public function testShowFilesTableOutputBelowThreshold(): void
+    {
+        $this->tester->run([
+            'cloverfile'   => self::$fixtures['valid'],
+            'threshold'    => 90,
+            '--show-files' => true,
+        ]);
+
+        self::assertEquals(self::$fixtures['valid'], $this->tester->getInput()->getArgument('cloverfile'));
+        self::assertSame(90, $this->tester->getInput()->getArgument('threshold'));
+
+        $eol = \PHP_EOL;
+
+        $expected = '----------------------------- --------------------------------- ---------- ' . $eol;
+        $expected .= '  File                          Covered Elements/Total Elements   Coverage  ' . $eol;
+        $expected .= ' ----------------------------- --------------------------------- ---------- ' . $eol;
+        $expected .= '  /tmp/Example/String.php       36/38                             94.74%    ' . $eol;
+        $expected .= '  /tmp/Example/StringList.php   20/24                             83.33%    ' . $eol;
+        $expected .= ' ----------------------------- --------------------------------- ---------- ' . $eol;
+        $expected .= '  Overall Totals                56/62                             89.04%    ' . $eol;
+        $expected .= ' ----------------------------- --------------------------------- ----------';
+
+        self::assertEquals($expected, trim($this->tester->getDisplay()));
+        self::assertSame(Command::FAILURE, $this->tester->getStatusCode());
+    }
+
+    public function testShowFilesTableOutputEmpty(): void
+    {
+        $this->tester->run([
+            'cloverfile'   => self::$fixtures['empty'],
+            'threshold'    => 90,
+            '--show-files' => true,
+        ]);
+
+        self::assertEquals(self::$fixtures['empty'], $this->tester->getInput()->getArgument('cloverfile'));
+        self::assertSame(90, $this->tester->getInput()->getArgument('threshold'));
+
+        self::assertSame('[ERROR] Insufficient data for calculation. Please add more code.', trim($this->tester->getDisplay()));
+        self::assertSame(Command::FAILURE, $this->tester->getStatusCode());
     }
 
     protected static function stripWhitespace(string $output): string